fix(w3d): start and end of W3dChunk are of type UInt

+++ w3d/src/main/kotlin/de/darkatra/bfme2/w3d/model/W3dChunk.kt
@@ -5,8 +5,8 @@ import com.google.common.io.CountingInputStream
 data class W3dChunk(
     val type: W3dChunkType,
     val payload: W3dPayload,
-    val start: ULong,
-    val end: ULong,
+    val start: UInt,
+    val end: UInt,
 ) {
 
     internal companion object {
